mechatrainer/requirements.txt
async-timeout==4.0.3
bleak==0.21.1
bottle==0.12.25
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
conan==1.62.0
distlib==0.3.4
ECPy==1.2.5
exceptiongroup==1.3.0
fasteners==0.19
filelock==3.6.0
future==0.18.3
hidapi==0.14.0
idna==3.6
iniconfig==2.1.0
Jinja2==3.1.3
ledgerblue==0.1.48
libusb1==3.1.0
MarkupSafe==2.1.4
ndeflib==0.3.3
nfcpy==1.0.4
node-semver==0.6.1
packaging==25.0
patch-ng==1.17.4
pbr==5.8.1
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.6.0
pluginbase==1.0.1
protobuf==4.25.2
pycryptodome==3.20.0
pycryptodomex==3.20.0
pyDes==2.0.1
pyelftools==0.30
Pygments==2.17.2
PyJWT==2.8.0
pyobjc-core==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-CoreBluetooth==9.2
pyobjc-framework-libdispatch==9.2
pyserial==3.5
pytest==8.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gnupg==0.5.2
python-u2flib-host==3.0.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
stevedore==3.5.0
tomli==2.3.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==1.26.18
virtualenv==20.14.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
websocket-client==1.7.0
